server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.6.1
numpy==1.26.2
opencv-python==4.8.1.78
face-recognition==1.3.0
scikit-learn==1.3.2
Pillow==10.1.0
python-dotenv==1.0.0